% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runTestApp.R
\name{runTestApp}
\alias{runTestApp}
\title{Launch test shiny Apps}
\usage{
runTestApp(name, ...)
}
\arguments{
\item{name}{The name of the app to run}

\item{...}{arguments to pass to shiny::runApp}
}
\description{
Launch test shiny Apps
}
\examples{
\dontrun{
library(testApps)

# running an app
  runTestApp("testApp01")

# to get list of names of shiny apps
  runTestApp()
# this will give error with all possible
# shiny names contained in the package.
}

}
